requirements.txt
imageio==2.9.0
nltk==3.6.3
numpy==1.20.3
Pillow==8.4.0
scikit_learn==1.0.1
termcolor==1.1.0
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.2
transformers==4.11.3
